+++ Frameworks/OakTextView/src/OakTextView.mm
@@ -461,7 +461,7 @@ private:
 	ng::layout_t* _layout;
 };
 
-@interface OakTextView () <NSTextInputClient, NSDraggingSource, NSIgnoreMisspelledWords, NSChangeSpelling, NSTextFieldDelegate, NSTouchBarDelegate, NSAccessibilityCustomRotorItemSearchDelegate, OakUserDefaultsObserver>
+@interface OakTextView () <NSTextInputClient, NSDraggingSource, NSIgnoreMisspelledWords, NSChangeSpelling, NSTextFieldDelegate, NSAccessibilityCustomRotorItemSearchDelegate, OakUserDefaultsObserver>
 {
 	std::shared_ptr<document_view_t> documentView;
 	ng::callback_t* callback;
@@ -2951,78 +2951,6 @@ static void update_menu_key_equivalents (NSMenu* menu, std::multimap<std::string
 }
 
 // ============================
-
-// =============
-// = Touch Bar =
-// =============
-
-static NSTouchBarItemIdentifier kOTVTouchBarCustomizationIdentifier          = @"com.macromates.TextMate.otv.customizationIdentifer";
-static NSTouchBarItemIdentifier kOTVTouchBarItemIdentifierNavigateBookmarks  = @"com.macromates.TextMate.otv.navigateBookmarks";
-static NSTouchBarItemIdentifier kOTVTouchBarItemIdentifierAddRemoveBookmark  = @"com.macromates.TextMate.otv.addRemoveBookmark";
-
-- (NSTouchBar*)makeTouchBar
-{
-	NSTouchBar* touchBar = [NSTouchBar new];
-	touchBar.delegate = self;
-	touchBar.defaultItemIdentifiers = @[ kOTVTouchBarItemIdentifierAddRemoveBookmark, kOTVTouchBarItemIdentifierNavigateBookmarks, ];
-	touchBar.customizationIdentifier = kOTVTouchBarCustomizationIdentifier;
-	touchBar.customizationAllowedItemIdentifiers = @[ kOTVTouchBarItemIdentifierAddRemoveBookmark, kOTVTouchBarItemIdentifierNavigateBookmarks, ];
-
-	return touchBar;
-}
-
-- (NSTouchBarItem*)touchBar:(NSTouchBar*)touchBar makeItemForIdentifier:(NSTouchBarItemIdentifier)identifier
-{
-	if([identifier isEqualToString:kOTVTouchBarItemIdentifierAddRemoveBookmark])
-	{
-		NSImage* bookmarkImage = [NSImage imageNamed:@"RemoveBookmarkTemplate" inSameBundleAsClass:[self class]];
-		bookmarkImage.accessibilityDescription = @"add or remove bookmark";
-
-		NSCustomTouchBarItem* bookmarkButtonTouchBarItem = [[NSCustomTouchBarItem alloc] initWithIdentifier:kOTVTouchBarItemIdentifierAddRemoveBookmark];
-		bookmarkButtonTouchBarItem.view = [NSButton buttonWithImage:bookmarkImage target:self action:@selector(toggleCurrentBookmark:)];
-		bookmarkButtonTouchBarItem.visibilityPriority = NSTouchBarItemPriorityHigh;
-		bookmarkButtonTouchBarItem.customizationLabel = @"Add/Remove Bookmark";
-
-		return bookmarkButtonTouchBarItem;
-	}
-	else if([identifier isEqualToString:kOTVTouchBarItemIdentifierNavigateBookmarks])
-	{
-		NSSegmentedControl* navigateMarkerSegmentedControl = [NSSegmentedControl new];
-		navigateMarkerSegmentedControl.segmentCount = 2;
-		navigateMarkerSegmentedControl.target       = self;
-		navigateMarkerSegmentedControl.action       = @selector(performNavigateBookmarksSegmentAction:);
-		navigateMarkerSegmentedControl.trackingMode = NSSegmentSwitchTrackingMomentary;
-		navigateMarkerSegmentedControl.segmentStyle = NSSegmentStyleSeparated;
-
-		NSImage* goUpImage = [NSImage imageNamed:NSImageNameTouchBarGoUpTemplate];
-		goUpImage.accessibilityDescription = @"previous bookmark";
-		NSImage* goDownImage = [NSImage imageNamed:NSImageNameTouchBarGoDownTemplate];
-		goDownImage.accessibilityDescription = @"next bookmark";
-
-		[navigateMarkerSegmentedControl setImage:goUpImage forSegment:0];
-		[navigateMarkerSegmentedControl setImage:goDownImage forSegment:1];
-
-		NSCustomTouchBarItem* markersTouchBarItem = [[NSCustomTouchBarItem alloc] initWithIdentifier:kOTVTouchBarItemIdentifierNavigateBookmarks];
-		markersTouchBarItem.view = navigateMarkerSegmentedControl;
-		markersTouchBarItem.visibilityPriority = NSTouchBarItemPriorityHigh;
-		markersTouchBarItem.customizationLabel = @"Previous/Next Bookmark";
-
-		return markersTouchBarItem;
-	}
-
-	return nil;
-}
-
-- (void)performNavigateBookmarksSegmentAction:(id)sender
-{
-	switch([sender selectedSegment])
-	{
-		case 0: [self goToPreviousBookmark:self]; break;
-		case 1: [self goToNextBookmark:self];     break;
-	}
-}
-
-// =============
 
 - (void)insertSnippetWithOptions:(NSDictionary*)someOptions // For Dialog popup
 {